orchay/pyproject.toml
[project]
name = "orchay"
version = "0.1.0"
description = "WezTerm 기반 Task 스케줄러"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "textual>=1.0",
    "rich>=14.0",
    "watchdog>=4.0",
    "pydantic>=2.0",
    "fastapi>=0.115",
    "uvicorn[standard]",
    "jinja2>=3.0",
    "markdown-it-py[plugins,linkify]>=3.0",
    "pygments>=2.17",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.24",
    "ruff>=0.5",
    "pyright>=1.1",
]

[project.scripts]
orchay = "orchay.cli:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchay"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "ANN", "B", "C4", "SIM"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
reportUnusedFunction = "warning"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
